Fix nosource

+++ telegram-cli.spec
@@ -8,7 +8,6 @@ Group:		Internet/Messaging
 License:	GPLv2
 URL:		https://github.com/vysheng/tg
 Source:	master.zip
-nosource:	0
 
 BuildRequires:	lua-devel, openssl-devel, libconfig-devel, readline-devel, wget
 #Requires:	wget
@@ -23,7 +22,7 @@ Telegram is an Open Source messaging platform for mobile, desktop focused on pri
 [ -d %{name} ] && rm -Rfv %{name}
 mkdir %{name}
 cd %{name}
-wget -O %{sources} https://github.com/vysheng/tg/archive/master.zip
+[ -f %{sources} ] && wget -O %{sources} https://github.com/vysheng/tg/archive/master.zip
 unzip %{sources}
 cd tg-master
 ./configure
